requirements.txt
boto3
PyYAML
requests
aws-requests-auth